fix(OMN-14700): move EnumFullSuiteReason to canonical enums/ layer
The class-naming CI gate (scripts/validate_class_naming.py) requires Enum*
classes not live in a model_*.py file. Relocate EnumFullSuiteReason to
src/omnibase_core/enums/enum_full_suite_reason.py (canonical enum home, correct
layering: models import enums) and rewire the model, node __init__, selector_core,
and the scripts/ci re-export to import from there. Same StrEnum values -> byte-for-
byte-identical serialization; 216 selector tests + 33 node tests stay green.
