+++ scripts/ci/test_selection_models.py
@@ -13,8 +13,8 @@ script.
 
 from __future__ import annotations
 
+from omnibase_core.enums.enum_full_suite_reason import EnumFullSuiteReason
 from omnibase_core.models.nodes.test_selector.model_test_selection import (
-    EnumFullSuiteReason,
     ModelTestSelection,
     ModuleName,
     TestPath,
